requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
google-genai==1.9.0
langdetect==1.0.9
prometheus-client==0.20.0
aiohttp==3.9.3
